main: move rule-related routines to capa.rules
ref #1821
